requirements.txt
pyrogram
TgCrypto
python-dotenv
google-generativeai
librosa
numpy
soundfile
uvloop